SWANDROID-1745: [L] Excessive CEC logical address polling
We only poll once in NexusCecService if the CEC message
received is a POLLING message.  This reduces the overall
polling by one-half and places the full responsibility of
CEC logical address POLLING to Android's HDMI Control Service.

As an optimisation, the CEC device type property is only read
in once at construction time.

refs #SWANDROID-1745

Change-Id: I75a05b176a4c8f5ee021b84cf2cdd13c1887bfc9