hwc: make surface tracking conditional.
we know that there will be problems logged until the
fences are properly implemented; also similar issues
will be seen on the cursor layer using the current
composition model (which we will switch back to hwc
once we fix that mode).

so, for those reasons, we meanwhile avoid tracking and
alarming the user by default.  we can still see the
logs by enabling the property.

refs #SWANDROID-1606

Change-Id: I52067a34af9ebe81955251dbe10b225114bfee42
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>