hwc: adding support for overscan position settings.
adding api to hwc binder for get/set of the overscan
setting, if an overscan setting is assigned, it will
be applied through to the nsc layers.

update the 'bcm-adjust-screen-offset' application to
make use of the hwc binder api as a placeholder and
demo on how such feature can be executed.

refs #SWANDROID-1598

Change-Id: I8ea5e0656d5b659b5a1a54202c159ddf8295fc84
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>