Resetting bypass keyslot when closing pid channels
Nexus currently does not keep track of the bypass keyslot changes
for a pid channel in order to clean them up if the pid channel is
closed.  This can result in data being inaccessible to GLR/unrestricted
memory if the GR2R slot was chosen and the HW pid channel is reallocated
to another user.  This is not a trivial nexus change so for now we
can keep track of this in the secure decoder and cleanup when the
pid channel is closed.

refs #SWANDROID-1716

Change-Id: I82dce13925ea4250b61ba86da70d09afc375c0bb