selinux: add support for flashing selinux enabled image.
the basic problem with the way the images are currently flashed
to the device is that the selinux contexts are lost when copying
from the nfs mount to the target.  the busybox utility used does
not support preserving contexts.

however, copying direcly through 'dd' would achieve preserving
context thus allowing to run selinux enabled mode.

updating the stage_target_bootimage script to be selinux aware
and allowing user to decide in which mode to flash the images.

refs #SWANDROID-1768

Change-Id: Ie3e9216f1665df219b84940d5eeb255ea495f49e
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>