Adding cache flush after graphics operations as well as before
Occasionally some pixel data mismatches are found if the graphic
buffer used by the Graphics2D hardware is immediately read by the
CPU.  The operations flush the surface memory before the operation,
which should be sufficient but it appears a second flush after
is required.  This resolves the issue until we can identify if there
is another issue at hand.

refs #SWANDROID-1486

Change-Id: I593105306c4039d7099a2e86db27a4ac98c18b60