Fix Broadcom TV input with the latest "Live Channels" app
Add workaround for setup activity: call TunerHAL.release() only in
the last session. This is to hide a race between the "setup" session
and "live" session. It's possible that the "live" session starts the
live decode before "setup" session exist. On exit the "setup" session
releases tuner thus stopping the live decode.

refs #SWANDROID-1794

Change-Id: I4a61d0e6d5c0c3983d25e7e9e0e0907535a3340b
Signed-off-by: Tomek Bury <tomek@broadcom.com>