Converting HWC to use a single NEXUS_SurfaceClient
This converts HWC to use only a single NEXUS_SurfaceClient internally and
manage a set of framebuffers.  It allows for much simpler resource managment
and tracking of gralloc buffers since they are used immediately and not
owned by NSC afterward.  It will also allow for removal of the YV12->422
conversion logic in gralloc since it can now directly use Graphics2D instead
of relying on NSC which only permitted raster-order surfaces.

refs #SWANDROID-1736

Change-Id: I7c5ed3be7f4885d0e4f8237e86aa6d30fd2cf2a6