Revert "Added persistent data block partition"
This reverts commit 8340cdf719329cdc7faf25a9ee872ff375c25f17.

The newer GTS doesn't complain if the persistent data block feature
is not enabled (optional). The current implementation is not
working with SELinux and would require duplicating of the policy
in order to specify the exact block device node for each profile
(as opposed to the generic /dev/block/by-name/pst link). We decided
to remove this optional feature for now.

refs #SWANDROID-1298

Change-Id: I6dd286c4c9ddf5fe63dbd7a5f1bb1f5468b7a8d8