nxserver: rework setup and bootup of nxserver.
no longer invoke nxserver from system call, instead we
are the nxserver and we invoke the libnxserver directly,
this offers a much greater flexibility in terms of server
configurability, no longer need command line parameters
in particular.

what else to note:

- there is a bit of magic bundled up with nxserver
command line parsing which should not happen but which
is present for now until cleanup on the nxserver side
takes place.

- since we are the server and the binder-ipc, things could
be well cleaned out (to do next).

- renamed from nexusinit to nxserver to clearly state the
reworks scope.

refs #SWANDROID-1723

Change-Id: Id754c6dd5e4e02c1ac8305c1b29c0a8c511440f9
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>