release: Update aosp fetch directories
Updated aosp fetch directory list to current.

Updated aosp patches to be generated based on the default revision
unless user specified a specific baseline to use.

Removed release script directory from the exclude list as it does get
used now as part of the unpack process.  The release script path in the
exclude list is outdated so it was actually not doing anything, so just
cleaning up here.

refs #SWANDROID-1541

Change-Id: I9b481319468f97f74949c255e975c58eca23d8aa