unpack: fix on automount systems
The unpack script does not work properly
on systems that automount when the USB
key is plugged in. Also the mount path
is incorrect for newer versions of Ubuntu.
Make scripts more robust to deal with
these problems, and output obvious error
messages so the user knows when their
unpack failed.

refs #SWANDROID-987

Change-Id: I82db756928744ec40919369a41718062be5b57d4