SWANDROID-1929: Do not spawn the StandbyMonitorThread unless explictly required
Changed the default behaviour to not spawn a standby monitor thread unless
the client requests a "standbyMonitorCallback" during the call to
"createClientContext()".  We now set the option "joinSettings.ignoreStandbyRequest"
during the call to "NxClient_Join()" to disable checking for the client
to acknowledge a change in standby state in nxserverlib.

This should help to resolve issue SWANDROID-1457 too.

refs #SWANDROID-1929

Change-Id: I091728481203ff97d0afac4c89aa5bcecf693e93