Correcting clipping calculations
Clipping calcluations were clipping based on the surface rectangle itself
and not based on the framebuffer rectangle.  Because of this, the
cursor layer could be muted.  The overscan/AdjustScreenOffset tool
will still heavily distort the cursor and produce other problems
because of the way it applies the values, that still needs to be
revisited.

refs #SWANDROID-1926

Change-Id: I2f738c45bd0194bc5a0a58dbc7d9dc3413c0385f