Moving overscan to Nexus Surface Clients
The overscan calculations will now be handled via NSC allowing uniform
scaling of the graphic display (including cursor), multimedia and
sideband layers.  Everything will be rendered at the native framebuffer
resolution, and the overscan is handled by telling NSC the window is
larger or smaller than it is which will result in the correct scaling.

refs #SWANDROID-1926

Change-Id: I177a8d6d0e03648f409a696a3da8867bd72c9d62