SWANDROID-1948: Refactor NexusIPCClient "createClientContext" to remove redundant info
Part 2 of 2 changes:
===================

Clients that do not want to register their interest in monitoring the standby
request can now simply call "createClientContext()" without any arguments.  The
name of the client is now automatically obtained from the "getClient()" method
of "NexusIPCClientFactory".  The process ID is also automatically established
during the call to "getClient()".  This cleans up a lot of the initialisation
code for obtain a client context.

The "createClientContext()" method of NexusService is now simplified and only
requires the name and PID of the client, rather than the full config structure
which has fields such as standbyMonitorCallback which are redundant in the
context of the server.

Both libhwcomposer and liblights want the display handle from the client
that opened/acquired them.  As this is now owned by our nxserver, there is no
need to pass in the client argument (as this would be NULL), so the code now
uses the cleaner "NEXUS_Platform_GetObjects()" function instead.

Part 1 of the changes updated the Widevine Classic L3 source code to make use
of the cleaner "createClientContext()" API. This changeset also includes the
pre-compiled libraries and Java archive files that are hosted in
"libsecurity/wvclassic".

refs #SWANDROID-1948

Change-Id: I6e27f4045a07406ba7b152f711b981fc77c2312a