SWANDROID-1987: [7252S] Blank screen after suspend/resume
There is a timing race condition in the code whereby if multiple IR
events are received before being processed by the "threadloop", then
a keyevent is not processed.  This can occur if the POWER button
on the Broadcom silver remote control is pressed and held down for
over 150ms.

The solution is not to process key repeats until the "threadloop" has
emitted a KEY PRESSED event.

This has been tested with l-release, and the STB can now exit S2
standby correctly.

refs #SWANDROID-1987

Change-Id: I331b802aa68d8d03311e49be7f6946eda5d7e4e6