Removing dynamic allocation when buffers are queued on a port
The buffer node structure was allocated once to link a buffer to the port
during AllocBuffer/UseBuffer.  Then, when Empty/Fill was called another
node structure was allocated to queue it on the port.  The second allocation
was being freed, but it's unnecessary dynamic allocation that was confusing
and can be removed.

refs #SWANDROID-2012

Change-Id: I141d387b34fd539f057cb4c9d2d3b8d33f4b9879