SWANDROID-1982: Prevent potential deadlock in standby when getHdmiOutputStatus() is called.
We now check to see whether we are in standby before making the underlying call
to "NxClient_GetDisplayStatus()".  This later function will deadlock if called in
standby.
This may potentially fix any deadlocks where the HDMI CEC code may call the
"getHdmiOutputStatus()" whilst in S1 standby.  The tracing of the code shows that
this is unlikely, but this is more of a defensive measure.

refs #SWANDROID-1982

Change-Id: Ide3f9a2fdd998fb90c02d3a26ed8dea558966bcf