SWANDROID-2080: [CEC] Consolidate checking for CEC Device Type
We now consolidate the checking for the CEC device type to two
locations.  One, the NexusService and two, NexusIPCClient.  The
former is likely to be deprecated in the future and is why I
decided not to make an actual binder call between the client
and server.

The getCecDeviceType now correctly returns "b_cecDeviceType" and
the check for an Invalid device type is used rather than the
magic number of "-1".

As the NexusIPCClient API has changed, the WV security libs have
also had to be updated to stay in sync.

refs # SWANDROID-2080

Change-Id: I4a1a57a49bb32f3de860370dee3cec94f0146aa3