SWANDROID-2097: [CEC] STB wakes up unexpectedly during transition to standby
Added extra check to see whether we are still powered up (state S0) to
decide whether or not to wake the STB up on reception of a CEC message
when the HDMI-CEC HAL believes it is in standby.

This avoids the race condition where the Power HAL has not placed the
STB in to standby before another CEC message from the TV is received.

This issue was easily reproduced on a Samsung 23 inch model where the
TV will broadcast multiple <Standby> CEC messages when the TV's
power button is pressed.

refs #SWANDROID-2097

Change-Id: I8fdc20b127a1e55c436bac570750892ba10240c3