SWANDROID-2084: [PM] Create an S0.5 standby state to only disconnect the HDMI output
Part 1 of 2 changes:
===================
S0.5 is not an official state but a "half-way house" state between S0 and S1.  It
allows the CPU and Nexus to be kept "running" but disconnects the HDMI output.  It
also places the CEC in a state whereby it can react to wake-up CEC messages from the
TV.

With this fix, it is now possible to have applications crash/terminate and restart
whilst in S0.5 without the system deadlocking (please see SWANDROID-1083) top-level
parent issue for more information about the other relevant changes needed.

refs #SWANDROID-2084

Change-Id: I29e4703b3b8ff1dcb34393d8bae88ce9c31eb45c