BOMX: Synchronize component thread termination
This introduces an event at the component thread termination to
synchronize with the OMX component destruction to make sure the thread
has finished its task before the component is freed. This fixes a race
condition when the video component destructor has completed while the
component thread is still trying to call into video decoder's virtual
functions.

refs #SWANDROID-2118

Change-Id: I59a5252eff2dbdfae9f483b86b98cd06f56857e4