SWANDROID-2136: [PM] Cannot wakeup from shutdown
Fixed the issue where at boot-up the wake sources were not being
passed to Nexus.  As a consequence, when the shutdown process
was initiated via the long-press of the POWER key, no wake-up
sources were already set and the STB could not wake back up.

Also, we now only enable/disable the HDMI output if it has changed
state (small optimisation).

refs #SWANDROID-2136

Change-Id: Ie7041425fb05e0dd53e34eee00a6dc18bf746ad1