Add real-time logger priority update and logger size configuration
This feature is useful when dumping large amount of information from
NEXUS while we don't want to impact on the scheduler for most of the
time. With this change, one can change the logger priority prior to
dumping the NEXUS logs and restore the priority back after the dump.
This also allows logger buffer size configuration.

refs #SWANDROID-2018

Change-Id: I9f279c938ecc83c7d22ab74fd0d26695a463fc74