gralloc/hwc: implicit managed access for palette.
to avoid leaking refcount, we want to force using 'managed
access'-alike for surfaces created out of existing memory.

effectively, we force the proper API use and explicit lock/unlock
behavior implementation.

'true' managed access isn't enabled in this delivery as it
appears to need more work to be actually doing what we need.

refs #SWANDROID-1861