SWANDROID-2153: [CEC] Avoid excessive frame re-transmission
Updated the "sendCecMessage()" method to allow the maximum number
of retries to be specified.  If the Android framework is sending
a CEC message, then the retry count is 0 to allow only the Android
framework to control how many retry attempts should be made.

If the internal HAL code (e.g. HDMI-CEC HAL) needs to directly
send a CEC message, then the maximum retry count defaults to 5
(maximum amount as specified in the HDMI/CEC 1.4b spec).

By providing this flexibility, we now avoid the situation whereby
both the Android framework *AND* NexusCecService are trying to
account for the number of CEC frame re-transmissions and
inadvertently retrying too many times.

refs #SWANDROID-2153

Change-Id: I4e587615381721685bd4c36ad5787328fab73c77