tools: add 'makehwcfg' to create default hwcfg img
The 'makehwcfg' tool is used to build a default empty hwcfg.img such that
it can be fastboot flashed to 'hwcfg' partition which is expected to have
cramfs.

The empty hwcfg.img should only be used when user does not have the appropriate
drm.bin to be used for creating the hwcfg.img and the hwcfg partition cannot
be formatted with cramfs (e.g. when we use platforms with eMMC and
only the gpt.bin can be flashed to create the partitions).

refs #SWANDROID-2109

Change-Id: I9ffee799709184331cd6e751b56531ce07829356
Signed-off-by: Wendy Ng <wendy.ng@broadcom.com>