SWANDROID-2195: [CEC] Physical address does not update after changing HDMI ports
The Nexus CEC module needs to get the CEC physical address from the HDMI output
nexus module.  We now pass this information to the CEC module during an HDMI
hotplug event.

The hdmi_cec HAL needs to re-read the CEC physical address whenever the HDMI
cable can be moved.  This is now catered for during hot-plug connected events
and during exit from standby.

There are still changes needed in refsw to ensure that the CEC physical address
is read back from the TV's EDID during the transition out of standby.  This
is captured in JIRA SWANDROID-2197.

Cleanup of the "#if NEXUS_HAS_CEC" conditional compilation directives will be
made under JIRA SWANDROID-2216.

refs #SWANDROID-2195

Change-Id: Icf859e25137110c3deac37c8b309bc1fb1b30800