release: change release script to packag refsw as an option
Previously, we always include REFSW (filtered) source code
in our releases, now we decide to release the patch (if any)
against an official URSR release and don't include refsw source
in the release tarball by default.  In order to compile the Android
release tarball, one must source the corresponding URSR release
that works with the Android release and preload it into the release
workspace. Please refer to readme.txt for more details.

refs #SWANDROID-2218

Change-Id: I851783c4a4843a6fee2b072836f95a5dbac9dcb5