SWANDROID-2242: HDMI CEC is re enabled at each boot
There is a bug in the Android HDMI/CEC framework whereby the
hdmi_cec HAL is always enabled at boot-time even if previously
disabled.  The Android property "persist.sys.hdmi.enable_cec"
was being set to "1" during the initialisation of the hdmi_cec
HAL.

The changes contained within this JIRA, break the link between
the Android HDMI/CEC framework and the HAL.  The same Android
property is used, but can now be cleared to "0" to disable CEC
HAL functionality across reboots.

The "setCecEnabled()" method is therefore no longer required to
be called by the hdmi-cec HAL upon initialisation and therefore
this NexusIPCClient API can be removed.  Because there is a change
in the NexusIPCClient API, the Widevine libs needs to be
regenerated to ensure compatibility.

refs #SWANDROID-2242

Change-Id: I392a6a2d1eead995bd2691fae496b38b45916fc2