Add HDMI hotplug detection switch logic
Adds logic to control "hdmi_audio" switch in the HDMI
hotplug detection callback. The switch state is controlled
by interacting with HDMI hotplug switch driver through
open/close/ioctl calls to "/dev/hdmi_hpd". The switch is
set to '1' when the HDMI port is connected and '0' when
the port is unplugged.

refs #SWANDROID-2294

Signed-off-by: Danesh Petigara <dpetigara@broadcom.com>