nexus-join: make use of the built in default for 'password'.
until now we had to specify/duplicate the code related to password
certificate join each time a nexus-join is done, now we fixed the
nxclient side to take care of parsing the password file if available,
that results in simplifyiung the api.

note: requires nexus update to work properly.

refs #SWANDROID-2264

Signed-off-by: Pierre Couillaud <pierre@broadcom.com>

Conflicts:
	bcm_platform/libnexusservice/nexusnxservice.cpp