gpt: support '-' flag for guessing start and size.
Use '-' as the start address if you want the partition to begin
immediately after the previous partition (start=prev_start+prev_size).

Use '-' as the size if you want the partition to fill the available
space until the start of the next partition or the end of the device.
This is only possible if the next partition doesn't use '-' for its
start flag.

Adding support for a subset of the standard block size suffixes:
k and K for kibibyte, M for mebibyte, G for gibibyte.

Moving the -o option to the beginning of the command line in Android.mk
to match the usage expectations (options first, partitions last).
This will allow config files to use "--" if needed to stop getopt from
looking into the rest of the command line (partitions definitions with
'-' symbols).

refs #SWANDROID-2323

Change-Id: I589ee4956f28fd0bd02b6e5564d27c6532f0747b