Revert "Revert "d-cma: default heap grow to 8m""
This reverts commit 511166bf8256fff5f6c5f4e6f5a989c33e06c6bb.

the minimum allocation size for the typical use case is 8m, this is
for a 1080p buffer, which the most common android use case for the ATV
integration.  larger allocation size is typically (and only at this time)
for 4K buffer which is the non common (ie certification, some non secure
playback) use case, therefore we allow the growth allocator to choose
to grow properly if need be for the odd use case and keep otherwise the
default, most memory use efficient, setup.

refs #SWANDROID-2361

Change-Id: I174f73e39907b982f800f3e0285c78c3df924126
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>