bomx/mem: advertise presence of video components to memory allocator.
in order to allow allocation balancing between heaps, we need to ensure
the video components are known to the allocator, currently nexus is not
able to provide this information, so we pass it through the actual video
component themselves (this will work fine for android originated video
components, but not for nexus sidebanding, which will need to be taken
care of elsewhere).

on the memory allocator side, we use that information to prevent using
the video memory pool if we believe a video component is likely to use
it.  we also use this information to practively move the already allocated
blocks for graphics from the video heap ahead of the video memory being
actually requested - thus the need to advertise presence of the video
component as soon as possible in the component creation.  note: the
proactive move of memory out of the video heap isn't part of this change
yet.

we use an ashmem block such that it gets tracked accordingly and would
clear itself up even on abnormal component termination if such happens.

refs #SWANDROID-2346

Change-Id: I13c2b12dc97df0366616642b4217cc7ef0bb1514