SWANDROID-2449: [PM] Always disconnect/connect HDMI output in PowerHAL
We now always disconnect the video outputs when entering any standby
state and reconnect them when resuming from standby.  The code to do
this has been removed from NexusNxClient and placed in the proper
PowerHAL library.  It is triggered off the "setInteractive()" API,
which is the correct point at which to enable/disable the display.

refs #SWANDROID-2449

Change-Id: I332af51426826a1be20008dfdc9aa306d4bc60e6