WV: adding wv_playback app
refs #SWANDROID-2470

This needs the following refsw commit to be picked up:

http://git-irv-03.irv.broadcom.com/cgit/baseline.git/commit/?id=5c75198e379ffc1f790ea3e4f9ead6da05f44894

This is the first version of wv_playback app for android to play
widevine encrypted file. Currently it only supports the files that
can be downloaded from http://dash-mse-test.appspot.com/media.html
(Widevine files are listed under "OOPS_CENC").

Change-Id: Idcd2373347a2dd4cfbf5fc133524b1f7e859689c