hwc: pipeline optimization for video playback
In an attempt to improve video playback performance
on platforms with lower memc bandwidth and/or slower
2d core, the following optimizations were implemented:

1) When possible, merge the fill and blit operations
performed by the hwc when doing overlays, into a
single 2d graphics operation.

2) Use Neon instructions to scan the overaly frame and
skip it if it's fully transparent. This is the typical
use case of youtube full-screen video playback.

Note: Both optimizations above are only applied for
video playback cases.

refs #SWANDROID-2376

Change-Id: Ifba451333cbf79f84a4f005488ebef4a6a50f333
Signed-off-by: Luis Navarro <navarro@broadcom.com>