hwc: restore duplicate frame drop behavior.
detection and dropping of duplicate frame from android was lost
when transitioning the composition model from multiple to single
nexus client.

restore the behavior but limit it to single layer, which is the
only case where we can guarantee safe display redundancy at this
time and which happens to be the only use case we care about for
now.

refs #SWANDROID-1736

Change-Id: Ib162e210fcff4a2cbe2850a3d06e95cda9b0edfc
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>