prebuild: fix boot-up issues coming from adbd
Update 'init.broadcomstb.usb.tcp.rc' such that the /dev/usb-ffs/adb won't be
created on platforms that must be using adb-over-tcp. This allows adbd to be
started without access the USB FunctionFS gadget driver and the real Gadget
Controller that do not really exist on platform that cannot do adb-over-usb.
At the same time, we won't need to enable the USB Dummy HCD controller driver
in kernel which lead to other problem on platforms that have a real USB Gadget
Controller.

refs #SWANDROID-2466

Change-Id: I6ceacacc9156dd058ee1054a059b5e40f99046fe
Signed-off-by: Wendy Ng <wendy.ng@broadcom.com>