hd/uhd: hook up logic to switch display into uhd if available.
add logic to hdmi hotplug callback to determine if we are
connected to a uhd display and if so select it as preferred
over hd.  assert it to nexus which applies the configuration
for graphics upscaling and video support.

the base geometry of the display as reported to the android
continues to be 1080 as there is no native application of uhd
yet, when that time comes, a hotplug listener would be installed
in the hwc instead and we would do the same logic there, plus
inform android about the display selected through the existing
api (which is hardcoded to 1080 for now).

refs #SWANDROID-2267

Change-Id: I6d634a1553ef51008ac98eeb7e305bd03ab4a3ec
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>