bomx_video_decoder: Fix EOS buffer leak
When doing back to back video playback with apps using
NuPlayer, the internal buffer used by the OMX component
to return the EOS message to the application is not
recycled properly. This causes that after several iterations
of back to back playback, we exhaust the interal list of
buffers and at some point we're unable to return decoded
frames to the application.
Applications which run into this problem will eventually
lose video on the display.

refs #SWANDROID-2492

Change-Id: I45b42df0d93091a1b9ce480f18411139d8d59d03
Signed-off-by: Luis Navarro <navarro@broadcom.com>