yv12: various tools for ease debugging.
- 'yv12torgba': this application is a simple command line which takes
input a yv12 captured buffer and produces a converted (and scaled) output
following both the hwc path (convert+scale) and the v3d path (convert
but no scale).  this is mainly to verify integrity of the conversion
operations.

- 'nxhdlcap': this application is a simple command line to capture content
of a handle identifed nexus memory block, it can be used to grab any block
and not necessarily gralloc's ones, but was designed for this initial purpose,
you can (as example) find the block handle to capture from looking at the
'gralloc logs' or the 'dumpsys SurfaceFlinger'.

- 'rgbaview': this application is a simple command line to display content
of a captured rgba buffer stored in a file.

refs #SWANDROID-2391