SWANDROID-2520: [PM] Premature wake-up caused by HDMI hotplug when entering standby
We now always check to see whether the received CEC message is one of a set of
valid "wake-up" messages before deciding whether to forward it on to the Android
framework and spoof an HDMI hot-plug event.

This prevents the situation where Nexus is in S2 or S3 standby, but the kernel
hasn't yet frozen the user-space/kernel-space processes and any received CEC
message can cause the STB to prematurely wake-up.

refs #SWANDROID-2520

Change-Id: I0119f4fe127802b9d8554f7c28eb38b9dbc936b5