SWANDROID-2510: [PM] brcm_audio HAL does not correctly handle transitioning to standby
We now correctly use the "NxClient_RegisterAcknowledgeStandby()" API when instantiating
the "StandbyMonitorThread" and the "NxClient_UnregisterAcknowledgeStandby()" when
the object is destroyed.  This allows for the brcm_audio HAL to separately acknowledge
when it is ready to enter standby even if it's in the same process as another NxClient.

This fixes the problem whereby another NxClient in the same process can acknowledge
the standby state and cause nxserver to believe it can enter standby, although the
brcm_audio HAL NxClient is not ready to transition to standby.

refs #SWANDROID-2510

Change-Id: Ie9203d37d96432908edab4884b1ea7bf9d37f879