SWANDROID-2423: [PM] Create kernel PM driver to handle STB PM states
Part 1 of 2 changes:
===================
We now introduce a brand new "droid_pm" kernel driver that is
responsible for handling all PM interaction at the kernel driver
level.  This driver incorporates the wakeup driver functions to
avoid the need for supporting multiple PM related kernel drivers.
It also supports interacting with the Power HAL to synchronise entry
to S2 and S3 standby modes.

Instead of having a fixed delay of 5s in the PowerHAL before deciding
when to enter standby, we now hook in to the kernel's PM notifier
framework to know when the kernel is about to suspend.  The Power HAL
code has been re-factored to poll for this "going to suspend"
notification and uses it to place Nexus in to S2 or S3 standby.
It then signals to the droid_pm kernel driver that it is ready to enter
standby and the droid_pm driver will allow the kernel to either enter
STR deep sleep (S3) or standby (S2).

For S0.5, S1 and S5 modes, we now immediately try to enter these
standby modes without waiting for the 5s timeout.

For S1, we use the Nexus standby monitoring to know when all Nexus
clients have acknowledged before entering standby.

For S0.5, we continue to blank the screen immediately but do not
enter any real standby mode.  We do however use pmlib to reduce
power consumption of kernel owned resources (e.g. reduce number of
CPU's to 1).

For S5 we continue to use the system libary function to execute
the "svc power shutdown" command to place the system in to power-off
mode when a short press of the power button is made.

The default power off mode has now been changed from S2 to S3 to
provide a much lower power consumption.

The "droid_pm.ko" driver is now loaded immediately before "nexus.ko"
so that both Nexus and the Power HAL can immediately open and use it.

Part 2 of the changes is for the "device/broadcom" directory to
allow building of the "droid_pm" kernel driver and to add SELinux
policies.

refs #SWANDROID-2423

Change-Id: Iec07f4c12fce346f5e17b028984e58d5446f844b