SWANDROID-2580: [PM] Cannot suspend to S3 due to HDMI holding a wake-lock
We now invalidate the "mHotplugConnected" state during HDMI CEC HAL
initialisation.  This allows the HDMI CEC HAL to forward an HDMI
hotplug "disconnected" event to the framework if a hotplug
"connected" event was received by the STB *before* the CEC HAL
has been instantiated.

This race condition was preventing the CEC HAL from forwarding the
"disconnected" event, as the CEC HAL believed there to be no change in
the connected state, as originally "mHotplugConnected" was initialised
to "HDMI_NOT_CONNECTED".

By allowing the "disconnected" hot-plug event to be forwarded to the
HDMI Android framework service, it will release an internal wake-lock
in "HdmiCecLocalDevicePlayback.java" that was preventing the kernel
from suspending.

The problem can easily be reproduced by ensuring that the STB has fully
booted to the desktop or launcher.  Then make sure that the TV's A/V
input selection is showing the STB connection with CEC enabled.  Then
press the POWER button on the TV's remote control to place the TV in
to standby.  The TV should then broadcast a CEC <Standby> message to
automatically place the connected STB in to standby.  The STB will
begin to enter standby, but the kernel fails to suspend because the
HDMI Playback device is still holding a wakelock.  The "disconnected"
hot-plug event is never forwarded to the HDMI Playback device for the
wake-lock to be released and for the kernel to then suspend.

refs #SWANDROID-2580

Change-Id: Ib02500ae7f33e28b91bb58062bb2c762dfefb09a