SWANDROID-2591: [PM] Hang during entry to S1 standby shortly after initial boot has completed.
If the user presses the POWER button during the Widevine DRM
plug-in initialisation "flurry" after the system has booted
to the desktop for the first time, then nexus can dead-lock
during "NEXUS_Platform_Uninit()" time.  This only occurs for
S1 standby as the other standby modes do not exercise this
issue.

This initial solution checks to make sure that the system is
not entering standby before issuing the call to
"NxClient_Uninit()".

refs #SWANDROID-2591

Change-Id: I628eb4f323e50035ecf7a6560fab506067144686