tools: fix to ensure gpt.bin gets re-built
Before the fix gpt.bin won't be re-built even if it is removed from the output
folder as long as makegpt is up-to-date because gpt.bin is created as a
post-install step of makegpt.  Now create a new make target for gpt.bin so the
makefile system can re-build gpt.bin accordingly.

refs #SWANDROID-2597

Change-Id: If16fcd56f577e663e19767f6b3a593232a2b5515
Signed-off-by: Wendy Ng <wendy.ng@broadcom.com>