nxserver: enable composition bypass when asked for.
composition bypass is a special mode that the surface compositor
is to run in which allows to bypass an extra blit composition cycle
when pushing surfaces to the display, effectively shortening the
time to display, relieving the memory and m2mc bandwidth and in the
end allowing to gain ~5fps on averrage.

the condition to allow such are however specific to a valid configuation
including:

- single memory controller device.
- single HD display enabled.
- single push client on the surface compositor stack.

the above requirements are typically met on 7250 class devices for
any android graphics only use case (ie not with sideband video).

refs #SWANDROID-2452