dipslay-size: switch from 1080p to/from UHD.
this property isn't quite clearly defined, however we know some third party
application such as YouTube do care about its value and uses that to switch
in/out UHD video streaming.  we reflect the state of the device through the
connected output in the porperty to direct the application (such as YouTube
to send us what makes most sense).

as this time we only toggle 1080p/UHD, the usage and definition of this
property is going to be clarified eventually by android, at which point we
would adopt a more precise support.

refs #SWANDROID-2346