Checking encoder output frame rate against bounds
If an application does not respect the limits advertised in media_codecs.xml
we need to reject their frame rate requests in SetParameter as opposed
to waiting for the encoder start to fail.  If the encoder start fails as
part of transitioning into the executing state the media framework gets
unhappy and can assert.

refs #SWANDROID-2694

Change-Id: I53a9ecb47827032dce47732ee9a78c7a57fb42d8