Selecting RF DRC for legacy AAC decoder
The settings and behavior of the legacy AAC and dolby pulse
decoders are different for DRC.  The legacy decoder doesn't
support a programmable target level like the pulse decoder
does so we need to program RF mode to hit a correct DRC
target.  RF mode introduces CTS failures for dolby pulse so
we instead use line with a -23dB target.

refs #SWANDROID-2733

Change-Id: Ic4698c14c07510900a85ccf4f3694700e2311760