hwc/nx_ashmem: account for unbalanced unlock/free due to race conditions.
because of the async nature of the hwc composition in the fence based
architecture, it is possible to be in a situation where a block is
released while hwc is composing with it.  detect such situation by marking
external refcount on the ashmem block and defer free to the time the block
is no longer in used.

refs #SWANDROID-2797

Change-Id: I1b6dd3fedfd94de59f2d1989f17cbc9575664846
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>