Adding option to disable metadata output and use adaptive playback
To rework the decoder buffering model and decouple display and
decode we will need to disable metadata output mode because
we won't have a defined point to return buffers to the decoder
in that mode.  Since in AdaptivePlayback each buffer is actually
tied 1:1 with an OMX buffer we will always be told when the
buffer is actually being freed.  Default still enables metadata,
but this can be overridden with media.brcm.vdec_enable_medata = 0.

refs #SWANDROID-1760

Change-Id: I060d75d5e255996532e754ee15481e1c3324d94f