SWANDROID-2862: AVKO failed CEC compliance test about address allocation issue
Fixed issue whereby the HDMI hotplug logic was not providing a "disconnect"
event in the case when successive hotplug "connect" events were received.

This is down to the way nexus only reports back that an event(s) has
occurred and it is up to the software above nexus to determine the
state and take appropriate action.

In this specific case, the HPD line was being toggled by the Quantum
Data test equipment in a short amount of time which resulted in the
HDMI hotplug logic only recognising this as a "connect" event,  rather
than a "disconnect" followed by a "connect" event.  Because of this
lack of a transition, the HDMI CEC HAL code did not reset the CEC
physical/logical address and hence a stale CEC physical/logical address
was reported back during the test(s).

refs #SWANDROID-2862

Change-Id: Ia6994bcc2944bdc25887332eef155ac3b71a465a