vendor: audio and video support for android native tunneling
*** initial revision.

android native tunneling: defined by android as the 'tunneling' mode,
the platform must support audio hardware offload track, it must
support omx-tunneling setup through android's extension, finally the
video window is associated with a sideband layer.  a/v sync is left up
to the platform to perform.

from the bcm tunneling architecture already in place (which continues
to live on), the main updates are:

- audio: we introduce a new 'direct-tunnel' output device.  the hardware
offload is not quite a direct output nor a standard output, it is a mix
of both down to the nexus resource integration; it also must implement a
few additional audio hal functions that only offload track require. for
those reasons, we choose to implement it as a separate module.

there is a quirk at the audio-flinger level whereby on boot especially
the direct output is opened followed by a direct-tunnel output one without
closing the prior direct output, this results in a nexus resource contention
which we choose to address by managing nexus resources free/alloc at the
hal level (because we do not allow nexus's server resource grab). eventually
both stream get closed off, thus leading to no drawback.

- video: we introduce support for a 'tunnel' component knowledge, we choose
to make this component distinct from the non tunnel one (this is permitted
by android integration).  doing so allows to better control the life
cycle of the component and its features without needing to worry about
the resources usage model.  when in tunnel mode, most of the output port
processing goes away since the port is never populated (while it is created
and enabled per android framework).  the video window geometry is derived
from the assocaited sideband layer through hwc-binder, this is similar to
the non-tunnel video case, but it operates on a different resource.

bonus: there is a memory saving associated with not needing to populate the
output port.

- 'a/v sync': the a/v sync is represented by the 'hardware sync' handle
which provides integration between an offload track and a sideband video. in
our integration, this is done through a 'stc-channel'.  the stc-channel is
owned by the audio hal (device/output) and is exchanged through the android
framework with the video component (bomx) during setup of this latter.

refs #SWANDROID-2689

Change-Id: Ic8c027cc0d3cbe7a303c842c06040e05c687e59c
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>